front: extract computeOccurrenceName() to separate file
We'll make use of this function from elsewhere in the next commit.
