+++ front/src/modules/trainschedule/components/Timetable/PacedTrain/hooks/useOccurrences.tsx
@@ -2,22 +2,10 @@ import { useMemo } from 'react';
 
 import dayjs from 'dayjs';
 
+import computeOccurrenceName from 'modules/trainschedule/helpers/computeOccurrenceName';
 import type { OccurrenceId } from 'reducers/osrdconf/types';
 
 import type { Occurrence, PacedTrainWithDetails } from '../../types';
-
-export const computeOccurrenceName = (pacedTrainName: string, index: number): string => {
-  const endByNumber = /\b\w+\s\d+$/;
-
-  if (endByNumber.test(pacedTrainName)) {
-    const endOfPacedTrainName = Number(pacedTrainName.split(' ').pop());
-    return `${pacedTrainName.replace(/\s\d+$/, '')} ${endOfPacedTrainName + 2 * index}`;
-  }
-  if (!Number.isNaN(+pacedTrainName)) {
-    return `${+pacedTrainName + 2 * index}`;
-  }
-  return `${pacedTrainName} ${2 * index + 1}`;
-};
 
 type OccurrencesState = {
   occurrences: Occurrence[];
